feat: 삭제 시 confirm 확인 메시지 추가

+++ src/main/resources/static/js/article.js
@@ -3,13 +3,16 @@ const deleteButton = document.getElementById('delete-btn');
 
 if (deleteButton) {
     deleteButton.addEventListener('click', ev => {
-        let id = document.getElementById('article-id').value;
-        fetch(`/api/articles/${id}`, {
-            method: 'DELETE'
-        })
-            .then(() => {
-                alert('삭제가 완료되었습니다.');
-                location.replace('/articles');
-            });
+        if (confirm("정말로 이 글을 삭제 하시겠습니까?")) {
+            let id = document.getElementById('article-id').value;
+            fetch(`/api/articles/${id}`, {
+                method: 'DELETE'
+            })
+                .then(() => {
+                    alert('삭제가 완료되었습니다.');
+                    location.replace('/articles');
+                });
+        }
+
     });
 }
